Delete newline after processing with sed
sed requires that the inputs are text files, and POSIX defines a
text file as zero or more lines, each with a trailing newline.

Some implementations of sed will always print a newline after the
line, even if it didn't originally have one, causing an extra blank
line when the newline is re-added below.