ssh-keygen: fix parsing of public exponent in ssh2 private keys
The fourth ignored u32 in an ssh2 private key is actually part of
the algorithm-specific key data.

For RSA, it is the bit-length of the private exponent, just as for
the following bignums d, n, iqmp, q, and p. So, we can just use the
same buffer_get_bignum_bits function to parse it. This fixes parsing
of e when its byte-length is not e[0] < 30 ? 3 : 1.

For DSA, it indicates whether or not the parameters are predefined.
If it is 0, the p, q, and g bignums follow. If it is 1 then it is
followed by string specifying some fixed parameters. Since ssh-keygen
only supports keys with p, q, and g given explicitly, check to make
sure it is 0.